+++ packages/expo-audio/src/Audio.types.ts
@@ -438,6 +438,10 @@ export type RecordingOptionsAndroid = {
    * `65536`
    */
   maxFileSize?: number;
+  /**
+   * The desired audio Source. See the [`AndroidAudioSource`](#androidaudiosource) enum for all valid values.
+   */
+  audioSource?: RecordingSource;
 };
 
 export type AudioMode = {
@@ -497,3 +501,23 @@ export type InterruptionMode = 'mixWithOthers' | 'doNotMix' | 'duckOthers';
  * @platform android
  */
 export type InterruptionModeAndroid = 'doNotMix' | 'duckOthers';
+
+/**
+ * Recording source for android.
+ *
+ * An audio source defines both a default physical source of audio signal, and a recording configuration.
+ * @see https://developer.android.com/reference/android/media/MediaRecorder.AudioSource
+ * @platform android
+ */
+export type RecordingSource =
+  | 'camcorder'
+  | 'default'
+  | 'mic'
+  | 'remote_submix'
+  | 'unprocessed'
+  | 'voice_call'
+  | 'voice_communication'
+  | 'voice_downlink'
+  | 'voice_performance'
+  | 'voice_recognition'
+  | 'voice_uplink';